fix button, router after login

+++ src/app/shared/navbar/navbar.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, ElementRef } from '@angular/core';
 import { ROUTES } from '../../sidebar/sidebar.component';
 import {Location, LocationStrategy, PathLocationStrategy} from '@angular/common';
 import {Router} from "@angular/router";
+import Swal from 'sweetalert2';
 
 @Component({
     // moduleId: module.id,
@@ -58,10 +59,29 @@ export class NavbarComponent implements OnInit{
     }
 
     logOut() {
-        localStorage.clear();
-        this.router.navigate(['home']).then(()=>{
-            location.reload()
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, Logout!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Deleted!',
+                    'Go to Home Page!',
+                    'success'
+                )
+                localStorage.clear();
+                this.router.navigate(['home']).then(()=>{
+                    location.reload()
+                })
+            }
         })
+
+
     }
 
 }
